+++ db/migrate/20150818195113_remove_first_and_last_answer_id_from_answer_sessions.rb
@@ -1,0 +1,15 @@
+class RemoveFirstAndLastAnswerIdFromAnswerSessions < ActiveRecord::Migration
+  def up
+    remove_index :answer_sessions, :first_answer_id
+    remove_column :answer_sessions, :first_answer_id
+    remove_index :answer_sessions, :last_answer_id
+    remove_column :answer_sessions, :last_answer_id
+  end
+
+  def down
+    add_column :answer_sessions, :last_answer_id, :integer
+    add_index :answer_sessions, :last_answer_id
+    add_column :answer_sessions, :first_answer_id, :integer
+    add_index :answer_sessions, :first_answer_id
+  end
+end